Handle already existing file in target with same name With this the PatternFileMover is able to run properly even if there is a same-naming file in the target directory. Before this change PatternFileMover simply does nothing if it stumples upon this case.
Attention: PatternFileMover will remove the already existing file without any warning! This will be improved in a further version by adding options to choose from what to do in this case.

Furthermore PatternFileMover will not stop working if a name association is broken. It skips the broken association from now on.